+++ tx-backend/src/integration/groovy/org/eclipse/tractusx/traceability/IntegrationSpecification.groovy
@@ -23,8 +23,8 @@ package org.eclipse.tractusx.traceability
 
 import com.xebialabs.restito.server.StubServer
 import groovy.json.JsonBuilder
-import org.eclipse.tractusx.traceability.assets.domain.asbuilt.AssetAsBuiltRepository
-import org.eclipse.tractusx.traceability.assets.domain.asplanned.AssetAsPlannedRepository
+import org.eclipse.tractusx.traceability.assets.domain.asbuilt.repository.AssetAsBuiltRepository
+import org.eclipse.tractusx.traceability.assets.domain.asplanned.repository.AssetAsPlannedRepository
 import org.eclipse.tractusx.traceability.assets.domain.base.BpnRepository
 import org.eclipse.tractusx.traceability.bpn.mapping.domain.ports.BpnEdcMappingRepository
 import org.eclipse.tractusx.traceability.common.config.ApplicationProfiles