+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/assets/application/rest/response/DescriptionsResponseTest.java
@@ -20,7 +20,9 @@
 package org.eclipse.tractusx.traceability.assets.application.rest.response;
 
 
-import org.eclipse.tractusx.traceability.assets.domain.model.Descriptions;
+import assets.response.DescriptionsResponse;
+import org.eclipse.tractusx.traceability.assets.application.asbuilt.mapper.AssetAsBuiltResponseMapper;
+import org.eclipse.tractusx.traceability.assets.domain.asbuilt.model.Descriptions;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -35,7 +37,7 @@ class DescriptionsResponseTest {
         final Descriptions response = new Descriptions(id, shortId);
 
         // when
-        final DescriptionsResponse result = DescriptionsResponse.from(response);
+        final DescriptionsResponse result = AssetAsBuiltResponseMapper.from(response);
 
         // then
         assertThat(result).usingRecursiveComparison()